feat: generateUserRPS 메서드선언
