+++ RockPaperScissors/RockPaperScissors/RPS.swift
@@ -1,0 +1,5 @@
+enum RPS {
+    case rock
+    case paper
+    case scissors
+}
